+++ shared/constants/login.js
@@ -1,17 +1,5 @@
 // @flow
-import HiddenString from '../util/hidden-string'
-
-export type Mode =
-  | 'codePageModeScanCode'
-  | 'codePageModeShowCode'
-  | 'codePageModeEnterText'
-  | 'codePageModeShowText'
-
-export type DeviceRole =
-  | 'codePageDeviceRoleExistingPhone'
-  | 'codePageDeviceRoleNewPhone'
-  | 'codePageDeviceRoleExistingComputer'
-  | 'codePageDeviceRoleNewComputer'
+import * as Types from './types/login'
 
 export const codePageDeviceRoleExistingPhone = 'codePageDeviceRoleExistingPhone'
 export const codePageDeviceRoleNewPhone = 'codePageDeviceRoleNewPhone'
@@ -26,10 +14,10 @@ export const codePageModeShowText = 'codePageModeShowText'
 export const countDownTime = 5 * 60
 
 export function defaultModeForDeviceRoles(
-  myDeviceRole: DeviceRole,
-  otherDeviceRole: DeviceRole,
+  myDeviceRole: Types.DeviceRole,
+  otherDeviceRole: Types.DeviceRole,
   brokenMode: boolean
-): ?Mode {
+): ?Types.Mode {
   switch (myDeviceRole + otherDeviceRole) {
     case codePageDeviceRoleExistingComputer + codePageDeviceRoleNewComputer:
       return codePageModeEnterText
@@ -64,30 +52,33 @@ export function qrGenerate(code: string): string {
   const qrCode = src.split('"')[1]
   return qrCode
 }
-// It's the b64 encoded value used to render the image
-type QRCode = HiddenString
 
-export type State = {
+export const initialState: Types.State = {
   codePage: {
-    cameraBrokenMode: boolean,
-    codeCountDown: number,
-    enterCodeErrorText: string,
-    mode: ?Mode,
-    myDeviceRole: ?DeviceRole,
-    otherDeviceRole: ?DeviceRole,
-    qrCode: ?QRCode,
-    qrCodeScanned: boolean,
-    qrScanned: ?QRCode,
-    textCode: ?HiddenString,
+    cameraBrokenMode: false,
+    codeCountDown: 0,
+    enterCodeErrorText: '',
+    mode: null,
+    myDeviceRole: null,
+    otherDeviceRole: null,
+    qrCode: null,
+    qrCodeScanned: false,
+    qrScanned: null,
+    textCode: null,
   },
-  configuredAccounts: ?Array<{|hasStoredSecret: boolean, username: string|}>,
-  forgotPasswordError: ?Error,
-  forgotPasswordSubmitting: boolean,
-  forgotPasswordSuccess: boolean,
-  justDeletedSelf: ?string,
-  justRevokedSelf: ?string,
-  loginError: ?string,
-  registerUserPassError: ?string,
-  registerUserPassLoading: boolean,
-  waitingForResponse: boolean,
+  configuredAccounts: null,
+  deviceName: {
+    deviceName: '',
+    existingDevices: [],
+    onSubmit: () => {},
+  },
+  forgotPasswordError: null,
+  forgotPasswordSubmitting: false,
+  forgotPasswordSuccess: false,
+  justDeletedSelf: null,
+  justRevokedSelf: null,
+  loginError: null,
+  registerUserPassError: null,
+  registerUserPassLoading: false,
+  waitingForResponse: false,
 }
